+++ database/migrations/2020_11_23_033618_create_transactions_table.php
@@ -15,12 +15,12 @@ class CreateTransactionsTable extends Migration
     {
         Schema::create('transactions', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('garage_id');
+            $table->unsignedInteger('garage_id');
             $table->time('start_time');
             $table->time('end_time');
             $table->string('total');
-            $table->unsignedBigInteger('car_in_by_user_id');
-            $table->unsignedBigInteger('car_out_by_user_id');
+            $table->unsignedInteger('car_in_by_user_id');
+            $table->unsignedInteger('car_out_by_user_id');
             $table->timestamps();
             $table->softDeletes();
         });
